+++ src/types/request_types.ts
@@ -1,0 +1,13 @@
+import { Request } from 'express';
+import { ParamsDictionary, Query } from 'express-serve-static-core';
+
+export type TxRequestBody = {
+	tx: string;
+};
+
+export type TxRequest = Request<
+	ParamsDictionary,
+	unknown,
+	TxRequestBody,
+	Query
+>;
